More stable parallel test runs

+++ setup.py
@@ -69,6 +69,9 @@ def render_templates_dir(mustache_hash, src, dst):
 def render_templates_leaf(mustache_hash, src, dst):
     assert src.exists(), "cannot use a nonexistent source"
     assert not src.is_dir(), "render_template_leaf called with a directory"
+    if dst.exists() and dst.stat().st_mtime > src.stat().st_mtime:
+        # the destination is newer. No need to do anything.
+        return
     with src.open(mode='r') as src_file:
         content = src_file.read()
     with dst.open(mode='w') as dst_file:
